% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{BT_color_picker}
\alias{BT_color_picker}
\title{Create a discrete color scale that can be plugged into other tools like Seurat.}
\usage{
BT_color_picker(n)
}
\description{
Create a discrete color scale that can be plugged into other tools like Seurat.
}
