% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_color_brian_c}
\alias{scale_color_brian_c}
\title{This function creates a continuous color palette.}
\usage{
scale_color_brian_c(
  ...,
  aesthetics = "color",
  color_palette = "Red",
  reverse = FALSE,
  Name = waiver()
)
}
\description{
This function creates a continuous color palette.
}
