% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_color_brian_d}
\alias{scale_color_brian_d}
\title{This function creates a discrete fill scale for up to 8 radical colors that can be plugged into any ggplot graph.}
\usage{
scale_color_brian_d(..., aesthetics = "color", Name = waiver())
}
\description{
This function creates a discrete fill scale for up to 8 radical colors that can be plugged into any ggplot graph.
}
