% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_fill_brian_c}
\alias{scale_fill_brian_c}
\title{This function creates a continuous fill palette.}
\usage{
scale_fill_brian_c(
  ...,
  aesthetics = "fill",
  color_palette = "Red",
  reverse = FALSE,
  Name = waiver()
)
}
\description{
This function creates a continuous fill palette.
}
