% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_fill_brian_d}
\alias{scale_fill_brian_d}
\title{This function creates a discrete fill scale from the color palette for up to 8 colors. It can be used
with all ggplot graphs.}
\usage{
scale_fill_brian_d(..., aesthetics = "fill", Name = waiver())
}
\description{
This function creates a discrete fill scale from the color palette for up to 8 colors. It can be used
with all ggplot graphs.
}
