% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_fill_brian_diverging}
\alias{scale_fill_brian_diverging}
\title{This function creates a diverging fill palette.}
\usage{
scale_fill_brian_diverging(
  ...,
  aesthetics = "fill",
  color_palette = "Red",
  reverse = FALSE,
  Name = waiver()
)
}
\description{
This function creates a diverging fill palette.
}
